015 registrando e excluindo chave bcb

+++ src/main/kotlin/br/com/zup/b1gvini/pix/grpc/remove/RemoveChavePixService.kt
@@ -1,10 +1,13 @@
 package br.com.zup.b1gvini.pix.grpc.remove
 
+import br.com.zup.b1gvini.clients.BCB
+import br.com.zup.b1gvini.clients.DeletePixKeyRequest
 import br.com.zup.b1gvini.clients.ItauERP
 import br.com.zup.b1gvini.compartilhado.exceptions.BadRequestException
 import br.com.zup.b1gvini.compartilhado.exceptions.NotFoundException
 import br.com.zup.b1gvini.pix.model.ChavePix
 import br.com.zup.b1gvini.pix.repository.ChavePixRepository
+import io.micronaut.http.HttpStatus
 import io.micronaut.validation.Validated
 import org.slf4j.LoggerFactory
 import javax.inject.Inject
@@ -16,7 +19,8 @@ import javax.validation.Valid
 @Singleton
 class RemoveChavePixService(
     @Inject private val repository: ChavePixRepository,
-    @Inject private val itauERP: ItauERP
+    @Inject private val itauERP: ItauERP,
+    @Inject private val bcb: BCB
 ) {
     val logger = LoggerFactory.getLogger(this::class.java)
 
@@ -37,6 +41,13 @@ class RemoveChavePixService(
 
         logger.info("verificando se o cliente é o possuidor da chave...")
         if (cpfConta != cpfDonoPix) throw BadRequestException("Dados inconsistente.")
+
+
+        val request = DeletePixKeyRequest(chavePix.chave)
+        logger.info("Estabelecendo conexao com BCB...")
+        val response = bcb.deletarChavePix(chavePix.chave,request)
+        if(response.status != HttpStatus.OK) throw BadRequestException("Erro ao deletar pix no BCB")
         repository.deleteById(chavePix.id!!)
+        logger.info("deletado com sucesso")
     }
 }
